add cli_test and refactor
